config 03/01

+++ N1224_Module2/src/ss5_overrice_overloading/exercise3/Discount.java
@@ -1,5 +1,7 @@
 package ss5_overrice_overloading.exercise3;
 
+import java.math.BigDecimal;
+
 public interface Discount {
-    void applyDiscount(double price);
+    void applyDiscount(BigDecimal price);
 }
